move TestKsqlRestApp to be in io.confluent.ksql.rest.server package so KsqlRestApplication build method may be package private.
