+++ ksql-cli/src/test/java/io/confluent/ksql/cli/BasicAuthFunctionalTest.java
@@ -26,8 +26,8 @@ import io.confluent.common.utils.IntegrationTest;
 import io.confluent.ksql.rest.client.KsqlRestClient;
 import io.confluent.ksql.rest.client.RestResponse;
 import io.confluent.ksql.rest.entity.ServerInfo;
+import io.confluent.ksql.rest.server.TestKsqlRestApp;
 import io.confluent.ksql.test.util.EmbeddedSingleNodeKafkaCluster;
-import io.confluent.ksql.test.util.TestKsqlRestApp;
 import io.confluent.rest.RestConfig;
 import java.io.IOException;
 import java.net.URI;
